+++ gs.mjs
@@ -47,8 +47,6 @@ Hooks.on('renderChatMessage', (app, html, data) => {
 		const targets = Array.from(game.user.targets);
 		const activeTarget = targets[0].document.actor.getActiveTokens()[0];
 		const armorScore = targets[0].document.actor.system.defenses.minion.armor;
-		console.log("... checking targets", targets, data);
-
 
 		// Setting up roll with weapon damage
 		let damageString = weapon.system.power;
@@ -83,6 +81,7 @@ Hooks.on('renderChatMessage', (app, html, data) => {
 		event.preventDefault();
 		const monster = game.actors.get(event.currentTarget.dataset.monsterid);
 		const defType = event.currentTarget.dataset.type;
+		const snipeSkill = event.currentTarget.dataset.snipeskill;
 		const activeToken = monster.getActiveTokens()[0];
 		let rollLabel = '', rollResult = '';
 		if(defType === 'dodge'){
@@ -93,19 +92,23 @@ Hooks.on('renderChatMessage', (app, html, data) => {
 			rollResult = game.i18n.localize('gs.dialog.block.blockTotal');
 		}
 		let value = event.currentTarget.dataset.value;
+		console.log('... checking Snipe skill', value, snipeSkill);
 		let chatMessage = `<div class="chat messageHeader grid grid-7col">
-			<img src='${activeToken.document.texture.src}'><h2 class="actorName grid-span-6">${activeToken.document.name}: ${rollLabel}</h2>
+		<img src='${activeToken.document.texture.src}'><h2 class="actorName grid-span-6">${activeToken.document.name}: ${rollLabel}</h2>
 		</div>`;
 		if(value.includes('d')){
+			if(snipeSkill < 0) value += ` + ${snipeSkill}`;
 			const roll = new Roll(value);
 			await roll.evaluate();
 			chatMessage += `<div class="armorDodgeScore specialRollChatMessage">${rollResult}: ${roll._total}</div>`;
+			if(snipeSkill < 0) chatMessage += `<div class="skillScore specialRollChatMessage">Snipe: ${snipeSkill}</div>`;
 			roll.toMessage({
 				speaker: { actor: monster },
 				flavor: chatMessage
 			});
 		}else{
-			chatMessage += `<div class="armorDodgeScore specialRollChatMessage">${rollResult}: ${value}</div>`;
+			chatMessage += `<div class="armorDodgeScore specialRollChatMessage">${rollResult}: ${parseInt(value, 10) + parseInt(snipeSkill,10)}</div>`;
+			if(snipeSkill < 0) chatMessage += `<div class="skillScore specialRollChatMessage">Snipe: ${snipeSkill}</div>`;
 			ChatMessage.create({
 				speaker: { actor: monster },
 				flavor: chatMessage
@@ -126,7 +129,6 @@ Hooks.on('renderChatMessage', (app, html, data) => {
 		let currentHP = token.document.actor.system.lifeForce.value;
 		const totalDmg = Math.max(currentHP - (dmg - armorScore), 0);
 
-		console.log("... dmg:", dmg, totalDmg, armorScore);
 		if(target.type === 'character')
 			await target.update({
 				'system.lifeForce.value': totalDmg
@@ -162,8 +164,17 @@ async function weaponMacroHotbarDrop(data, slot){
 
         if (actor && item) {
             const command = `const actor = game.actors.get("${actor.id}");
-				if (actor.sheet) actor.sheet._newPlayerAttack("${item.id}");
-				else ui.notifications.warn("No active sheet for actor.");`;
+				const item = actor.items.get("${item.id}");
+				// Mock event object
+				const mockEvent = {
+					preventDefault: () => {},
+					currentTarget: {
+						dataset: {
+							itemid: "${item.id}"
+						}
+					}
+				};
+				actor.sheet._newPlayerAttack(mockEvent);`;
 
 			let macro = game.macros.find(m => (m.name === item.name) && (m.command === command));